+++ packages/nuomi/types/hooks.d.ts
@@ -2,6 +2,6 @@ import { Props } from './props';
 import { ConnectGetState } from './connect';
 import { NuomiStoreDispatch } from './store';
 
-export function useConnect<S = any>(getState?: ConnectGetState): [S, NuomiStoreDispatch];
+export function useConnect<S = any, G = any>(getState?: ConnectGetState<S, G>): [S & { getter: G }, NuomiStoreDispatch];
 
-export function useNuomi<S = any>(): [Props<S>];
+export function useNuomi<S = any, G = any>(): [Props<S, G>];