+++ packages/nuomi/types/nuomi.d.ts
@@ -1,6 +1,6 @@
 import { DefineProps } from './props';
 
-export type DefinePropsFunc = <S = any>(props: DefineProps<S>) => DefineProps<S>;
+export type DefinePropsFunc = <S = any, G = any>(props: DefineProps<S, G>) => DefineProps<S, G>;
 export interface ConfigureFunc {
   (): DefineProps;
   (props: DefineProps): DefineProps;