+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - reducers更名为reducer，移除默认reducer中的_updatState/_updateLoading/_replaceState，新增@update/@replace/@loading
   - effects更名为action，定义的方法接受2个参数，第一个是store，第二个是payload
   - 新增extends属性，类型为数组，可以拓展多个，取代0.x的nuomi.extend
+  - 新增getter属性，用于过滤或者组合state值
   - store对象新增commit和restoreState方法，dispatch用法更改，第一个参数为type，第二个参数为payload
   - state中的loadings对象更名为loading
 * Route组件